# luci-app-insomclash/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk aplikasi tunneling menggunakan mihomo",
        "title": "FusionTunX API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/app/config": {
            "get": {
                "description": "Get current FusionTunX application configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Get application configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update FusionTunX application configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Update application configuration",
                "parameters": [
                    {
                        "description": "Configuration to update",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/app/geo/ipv4": {
            "get": {
                "description": "Get IPv4 address with geolocation information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Get IPv4 geolocation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/app/geo/ipv6": {
            "get": {
                "description": "Get IPv6 address with geolocation information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Get IPv6 geolocation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/app/ipv4": {
            "get": {
                "description": "Get current IPv4 address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Get IPv4 address",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/app/ipv6": {
            "get": {
                "description": "Get current IPv6 address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FusionTunX App"
                ],
                "summary": "Get IPv6 address",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/backup/create": {
            "post": {
                "description": "Create a tar.gz backup of the entire working directory",
                "produces": [
                    "application/gzip"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Create backup",
                "responses": {
                    "200": {
                        "description": "Backup file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/backup/restore": {
            "post": {
                "description": "Restore configuration from uploaded tar.gz backup file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Restore backup",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Backup file (.tar.gz)",
                        "name": "backup",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dns/lookup": {
            "post": {
                "description": "Resolve domain name to IPv4 and IPv6 addresses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DNS"
                ],
                "summary": "Lookup domain IP addresses",
                "parameters": [
                    {
                        "description": "Domain to lookup",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LookupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LookupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.LookupResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.LookupResponse"
                        }
                    }
                }
            }
        },
        "/mihomo/active-config": {
            "get": {
                "description": "Get the currently active mihomo configuration file path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo files"
                ],
                "summary": "Get active config path",
                "responses": {
                    "200": {
                        "description": "Active config path",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Set the active mihomo configuration file path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo files"
                ],
                "summary": "Set active config path",
                "parameters": [
                    {
                        "description": "Filename to set as active",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/mihomo/dashboard-info": {
            "get": {
                "description": "Get external-controller port, secret, and available dashboards from ui directory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo"
                ],
                "summary": "Get Mihomo dashboard information",
                "responses": {
                    "200": {
                        "description": "Dashboard information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mihomo/restart": {
            "post": {
                "description": "Restart the mihomo service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo"
                ],
                "summary": "Restart mihomo service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mihomo/start": {
            "post": {
                "description": "Start the mihomo service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo"
                ],
                "summary": "Start mihomo service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/mihomo/stop": {
            "post": {
                "description": "Stop the mihomo service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mihomo"
                ],
                "summary": "Stop mihomo service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.LookupRequest": {
            "type": "object",
            "required": [
                "domain"
            ],
            "properties": {
                "domain": {
                    "type": "string"
                }
            }
        },
        "handler.LookupResponse": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "ipv4": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ipv6": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}